@@ -16,10 +16,11 @@ import (
 	"github.com/forceu/gokapi/internal/helper"
 	"github.com/forceu/gokapi/internal/logging"
 	"github.com/forceu/gokapi/internal/models"
+	"github.com/forceu/gokapi/internal/storage/chunking"
 	"github.com/forceu/gokapi/internal/storage/cloudstorage/aws"
 	"github.com/forceu/gokapi/internal/webserver/downloadstatus"
+	"github.com/jinzhu/copier"
 	"io"
-	"io/ioutil"
 	"log"
 	"mime/multipart"
 	"net/http"
@@ -30,41 +31,29 @@ import (
 	"time"
 )
 
-// NewFile creates a new file in the system. Called after an upload has been completed. If a file with the same sha256 hash
+// ErrorFileTooLarge is an error that is called when a file larger than the set maximum is uploaded
+var ErrorFileTooLarge = errors.New("upload limit exceeded")
+
+// NewFile creates a new file in the system. Called after an upload from the API has been completed. If a file with the same sha1 hash
 // already exists, it is deduplicated. This function gathers information about the file, creates an ID and saves
-// it into the global configuration.
+// it into the global configuration. It is now only used by the API, the web UI uses NewFileFromChunk
 func NewFile(fileContent io.Reader, fileHeader *multipart.FileHeader, uploadRequest models.UploadRequest) (models.File, error) {
 	if fileHeader.Size > int64(configuration.Get().MaxFileSizeMB)*1024*1024 {
-		return models.File{}, errors.New("upload limit exceeded")
+		return models.File{}, ErrorFileTooLarge
 	}
 	var hasBeenRenamed bool
-	reader, hash, tempFile, encInfo := generateHash(fileContent, fileHeader, uploadRequest)
+	reader, hash, tempFile, encInfo := generateHashAndEncrypt(fileContent, fileHeader)
 	defer deleteTempFile(tempFile, &hasBeenRenamed)
-	id := helper.GenerateRandomString(configuration.Get().LengthId)
-	file := models.File{
-		Id:                 id,
-		Encryption:         encInfo,
-		Name:               fileHeader.Filename,
-		SHA256:             hex.EncodeToString(hash),
-		Size:               helper.ByteCountSI(fileHeader.Size),
-		ExpireAt:           uploadRequest.ExpiryTimestamp,
-		ExpireAtString:     time.Unix(uploadRequest.ExpiryTimestamp, 0).Format("2006-01-02 15:04"),
-		DownloadsRemaining: uploadRequest.AllowedDownloads,
-		PasswordHash:       configuration.HashPassword(uploadRequest.Password, true),
-		ContentType:        fileHeader.Header.Get("Content-Type"),
-		UnlimitedTime:      uploadRequest.UnlimitedTime,
-		UnlimitedDownloads: uploadRequest.UnlimitedDownload,
+	header, err := chunking.ParseMultipartHeader(fileHeader)
+	if err != nil {
+		return models.File{}, err
 	}
-	if aws.IsAvailable() {
-		if !configuration.Get().PicturesAlwaysLocal || !isPictureFile(file.Name) {
-			aws.AddBucketName(&file)
-		}
-	}
-	addHotlink(&file)
-	filename := configuration.Get().DataDir + "/" + file.SHA256
+	file := createNewMetaData(hex.EncodeToString(hash), header, uploadRequest)
+	file.Encryption = encInfo
+	filename := configuration.Get().DataDir + "/" + file.SHA1
 	dataDir := configuration.Get().DataDir
 
-	if file.AwsBucket != "" {
+	if !file.IsLocalStorage() {
 		exists, size, err := aws.FileExists(file)
 		if err != nil {
 			return models.File{}, err
@@ -82,9 +71,9 @@ func NewFile(fileContent io.Reader, fileHeader *multipart.FileHeader, uploadRequ
 	fileWithHashExists := FileExists(file, configuration.Get().DataDir)
 	if fileWithHashExists {
 		encryptionLevel := configuration.Get().Encryption.Level
-		previousEncryption, ok := getEncInfoFromExistingFile(file.SHA256)
+		previousEncryption, ok := getEncInfoFromExistingFile(file.SHA1)
 		if !ok && encryptionLevel != encryption.NoEncryption && encryptionLevel != encryption.EndToEndEncryption {
-			err := os.Remove(dataDir + "/" + file.SHA256)
+			err = os.Remove(dataDir + "/" + file.SHA1)
 			helper.Check(err)
 			fileWithHashExists = false
 		} else {
@@ -94,15 +83,15 @@ func NewFile(fileContent io.Reader, fileHeader *multipart.FileHeader, uploadRequ
 
 	if !fileWithHashExists {
 		if tempFile != nil {
-			err := tempFile.Close()
+			err = tempFile.Close()
 			helper.Check(err)
-			err = os.Rename(tempFile.Name(), dataDir+"/"+file.SHA256)
+			err = os.Rename(tempFile.Name(), dataDir+"/"+file.SHA1)
 			helper.Check(err)
 			hasBeenRenamed = true
 			database.SaveMetaData(file)
 			return file, nil
 		}
-		destinationFile, err := os.OpenFile(filename, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0644)
+		destinationFile, err := os.OpenFile(filename, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
 		if err != nil {
 			return models.File{}, err
 		}
@@ -116,6 +105,194 @@ func NewFile(fileContent io.Reader, fileHeader *multipart.FileHeader, uploadRequ
 	return file, nil
 }
 
+func validateChunkInfo(file *os.File, fileHeader chunking.FileHeader) error {
+	maxFileSizeB := int64(configuration.Get().MaxFileSizeMB) * 1024 * 1024
+	if fileHeader.Size > maxFileSizeB {
+		return ErrorFileTooLarge
+	}
+	size, err := helper.GetFileSize(file)
+	if err != nil {
+		return err
+	}
+	if size != fileHeader.Size {
+		return errors.New("total filesize does not match")
+	}
+	return nil
+}
+
+// NewFileFromChunk creates a new file in the system after a chunk upload has fully completed. If a file with the same sha1 hash
+// already exists, it is deduplicated. This function gathers information about the file, creates an ID and saves
+// it into the global configuration.
+func NewFileFromChunk(chunkId string, fileHeader chunking.FileHeader, uploadRequest models.UploadRequest) (models.File, error) {
+	if chunkId == "" {
+		return models.File{}, errors.New("empty chunk id provided")
+	}
+	if !helper.FileExists(configuration.Get().DataDir + "/chunk-" + chunkId) {
+		return models.File{}, errors.New("chunk file does not exist")
+	}
+	file, err := chunking.GetFileByChunkId(chunkId)
+	if err != nil {
+		return models.File{}, err
+	}
+	defer file.Close()
+	err = validateChunkInfo(file, fileHeader)
+	if err != nil {
+		return models.File{}, err
+	}
+
+	var hash string
+	if uploadRequest.IsEndToEndEncrypted {
+		hash = "e2e-" + helper.GenerateRandomString(20)
+	} else {
+		hash, err = hashFile(file, isEncryptionRequested())
+		if err != nil {
+			_ = file.Close()
+			return models.File{}, err
+		}
+	}
+
+	metaData := createNewMetaData(hash, fileHeader, uploadRequest)
+
+	fileExists := FileExists(metaData, configuration.Get().DataDir)
+	if fileExists {
+		encryptionLevel := configuration.Get().Encryption.Level
+		previousEncryption, ok := getEncInfoFromExistingFile(metaData.SHA1)
+		if !ok && encryptionLevel != encryption.NoEncryption && encryptionLevel != encryption.EndToEndEncryption {
+			err = os.Remove(configuration.Get().DataDir + "/" + metaData.SHA1)
+			helper.Check(err)
+			fileExists = false
+		} else {
+			metaData.Encryption = previousEncryption
+		}
+	}
+
+	if fileExists {
+		err = file.Close()
+		if err != nil {
+			return models.File{}, err
+		}
+		err = os.Remove(file.Name())
+		if err != nil {
+			return models.File{}, err
+		}
+	}
+	if !isEncryptionRequested() {
+		if !fileExists {
+			_, err = file.Seek(0, io.SeekStart)
+			if err != nil {
+				return models.File{}, err
+			}
+			if !metaData.IsLocalStorage() {
+				_, err = aws.Upload(file, metaData)
+				if err != nil {
+					return models.File{}, err
+				}
+				database.SaveMetaData(metaData)
+				err = os.Remove(file.Name())
+				helper.Check(err)
+				return metaData, nil
+			}
+			err = os.Rename(file.Name(), configuration.Get().DataDir+"/"+metaData.SHA1)
+			if err != nil {
+				return models.File{}, err
+			}
+		}
+		database.SaveMetaData(metaData)
+		return metaData, nil
+	}
+	if !fileExists {
+		tempFile, err := encryptChunkFile(file, &metaData)
+		defer func() {
+			_ = file.Close()
+			_ = os.Remove(file.Name())
+			_ = tempFile.Close()
+			_ = os.Remove(tempFile.Name())
+		}()
+		if err != nil {
+			return models.File{}, err
+		}
+		if !metaData.IsLocalStorage() {
+			_, err = aws.Upload(tempFile, metaData)
+			if err != nil {
+				return models.File{}, err
+			}
+			tempFile.Close()
+			database.SaveMetaData(metaData)
+			return metaData, nil
+		}
+		tempFile.Close()
+		err = os.Rename(tempFile.Name(), configuration.Get().DataDir+"/"+metaData.SHA1)
+		if err != nil {
+			return models.File{}, err
+		}
+	}
+
+	database.SaveMetaData(metaData)
+	return metaData, nil
+}
+
+func encryptChunkFile(file *os.File, metadata *models.File) (*os.File, error) {
+	_, err := file.Seek(0, io.SeekStart)
+	if err != nil {
+		return nil, err
+	}
+	tempFileEnc, err := os.CreateTemp(configuration.Get().DataDir, "upload")
+	if err != nil {
+		return nil, err
+	}
+	encInfo := metadata.Encryption
+	err = encryption.Encrypt(&encInfo, file, tempFileEnc)
+	if err != nil {
+		return nil, err
+	}
+	_, err = tempFileEnc.Seek(0, io.SeekStart)
+	if err != nil {
+		return nil, err
+	}
+	metadata.Encryption = encInfo
+	err = file.Close()
+	if err != nil {
+		return nil, err
+	}
+	err = os.Remove(file.Name())
+	if err != nil {
+		return nil, err
+	}
+	return tempFileEnc, nil
+}
+
+func createNewMetaData(hash string, fileHeader chunking.FileHeader, uploadRequest models.UploadRequest) models.File {
+	file := models.File{
+		Id:                 createNewId(),
+		Name:               fileHeader.Filename,
+		SHA1:               hash,
+		Size:               helper.ByteCountSI(fileHeader.Size),
+		SizeBytes:          fileHeader.Size,
+		ContentType:        fileHeader.ContentType,
+		ExpireAt:           uploadRequest.ExpiryTimestamp,
+		ExpireAtString:     time.Unix(uploadRequest.ExpiryTimestamp, 0).Format("2006-01-02 15:04"),
+		DownloadsRemaining: uploadRequest.AllowedDownloads,
+		UnlimitedTime:      uploadRequest.UnlimitedTime,
+		UnlimitedDownloads: uploadRequest.UnlimitedDownload,
+		PasswordHash:       configuration.HashPassword(uploadRequest.Password, true),
+	}
+	if uploadRequest.IsEndToEndEncrypted {
+		file.Encryption = models.EncryptionInfo{IsEndToEndEncrypted: true, IsEncrypted: true}
+		file.Size = helper.ByteCountSI(uploadRequest.RealSize)
+	}
+	if aws.IsAvailable() {
+		if !configuration.Get().PicturesAlwaysLocal || !isPictureFile(file.Name) {
+			aws.AddBucketName(&file)
+		}
+	}
+	addHotlink(&file)
+	return file
+}
+
+func createNewId() string {
+	return helper.GenerateRandomString(configuration.Get().LengthId)
+}
+
 func getEncInfoFromExistingFile(hash string) (models.EncryptionInfo, bool) {
 	encryptionLevel := configuration.Get().Encryption.Level
 	if encryptionLevel == encryption.NoEncryption || encryptionLevel == encryption.EndToEndEncryption {
@@ -123,7 +300,7 @@ func getEncInfoFromExistingFile(hash string) (models.EncryptionInfo, bool) {
 	}
 	allFiles := database.GetAllMetadata()
 	for _, existingFile := range allFiles {
-		if existingFile.SHA256 == hash {
+		if existingFile.SHA1 == hash {
 			return existingFile.Encryption, true
 		}
 	}
@@ -139,6 +316,54 @@ func deleteTempFile(file *os.File, hasBeenRenamed *bool) {
 	}
 }
 
+const (
+	// ParamExpiry is a bit to indicate that the time remaining shall be changed after a duplication
+	ParamExpiry int = 1 << iota
+	// ParamDownloads is a bit to indicate that the downloads remaining shall be changed after a duplication
+	ParamDownloads
+	// ParamPassword is a bit to indicate that the password shall be changed after a duplication
+	ParamPassword
+	// ParamName is a bit to indicate that the filename shall be changed after a duplication
+	ParamName
+)
+
+// DuplicateFile creates a copy of an existing file with new parameters
+func DuplicateFile(file models.File, parametersToChange int, newFileName string, fileParameters models.UploadRequest) (models.File, error) {
+	var newFile models.File
+	err := copier.Copy(&newFile, &file)
+	if err != nil {
+		return models.File{}, err
+	}
+
+	changeExpiry := parametersToChange&ParamExpiry != 0
+	changeDownloads := parametersToChange&ParamDownloads != 0
+	changePassword := parametersToChange&ParamPassword != 0
+	changeName := parametersToChange&ParamName != 0
+
+	if changeExpiry {
+		newFile.ExpireAt = fileParameters.ExpiryTimestamp
+		newFile.ExpireAtString = time.Unix(fileParameters.ExpiryTimestamp, 0).Format("2006-01-02 15:04")
+		newFile.UnlimitedTime = fileParameters.UnlimitedTime
+	}
+	if changeDownloads {
+		newFile.DownloadsRemaining = fileParameters.AllowedDownloads
+		newFile.UnlimitedDownloads = fileParameters.UnlimitedDownload
+	}
+	if changePassword {
+		newFile.PasswordHash = configuration.HashPassword(fileParameters.Password, true)
+	}
+	if changeName {
+		newFile.Name = newFileName
+	}
+
+	newFile.Id = createNewId()
+	newFile.DownloadCount = 0
+	addHotlink(&file)
+
+	database.SaveMetaData(newFile)
+	return newFile, nil
+}
+
 // DeleteAllEncrypted marks all encrypted files for deletion on next cleanup
 func DeleteAllEncrypted() {
 	files := database.GetAllMetadata()
@@ -149,13 +374,25 @@ func DeleteAllEncrypted() {
 	}
 }
 
+func hashFile(input io.Reader, useSalt bool) (string, error) {
+	hash := sha1.New()
+	_, err := io.Copy(hash, input)
+	if err != nil {
+		return "", err
+	}
+	if useSalt {
+		hash.Write([]byte(configuration.Get().Authentication.SaltFiles))
+	}
+	return hex.EncodeToString(hash.Sum(nil)), nil
+}
+
 // Generates the SHA1 hash of an uploaded file and returns a reader for the file, the hash and if a temporary file was created the
 // reference to that file.
-func generateHash(fileContent io.Reader, fileHeader *multipart.FileHeader, uploadRequest models.UploadRequest) (io.Reader, []byte, *os.File, models.EncryptionInfo) {
+func generateHashAndEncrypt(fileContent io.Reader, fileHeader *multipart.FileHeader) (io.Reader, []byte, *os.File, models.EncryptionInfo) {
 	hash := sha1.New()
 	encInfo := models.EncryptionInfo{}
-	if fileHeader.Size <= int64(uploadRequest.MaxMemory)*1024*1024 {
-		content, err := ioutil.ReadAll(fileContent)
+	if fileHeader.Size <= int64(configuration.Get().MaxMemory)*1024*1024 {
+		content, err := io.ReadAll(fileContent)
 		helper.Check(err)
 		hash.Write(content)
 		if isEncryptionRequested() {
@@ -167,7 +404,7 @@ func generateHash(fileContent io.Reader, fileHeader *multipart.FileHeader, uploa
 		}
 		return bytes.NewReader(content), hash.Sum(nil), nil, encInfo
 	}
-	tempFile, err := os.CreateTemp(uploadRequest.DataDir, "upload")
+	tempFile, err := os.CreateTemp(configuration.Get().DataDir, "upload")
 	helper.Check(err)
 	var multiWriter io.Writer
 
@@ -178,7 +415,7 @@ func generateHash(fileContent io.Reader, fileHeader *multipart.FileHeader, uploa
 	helper.Check(err)
 
 	if isEncryptionRequested() {
-		tempFileEnc, err := os.CreateTemp(uploadRequest.DataDir, "upload")
+		tempFileEnc, err := os.CreateTemp(configuration.Get().DataDir, "upload")
 		helper.Check(err)
 		encryption.Encrypt(&encInfo, tempFile, tempFileEnc)
 		err = os.Remove(tempFile.Name())
@@ -202,13 +439,15 @@ func isEncryptionRequested() bool {
 		fallthrough
 	case encryption.FullEncryptionInput:
 		return true
+	case encryption.EndToEndEncryption:
+		return false
 	default:
 		log.Fatalln("Unknown encryption level requested")
 		return false
 	}
 }
 
-var imageFileExtensions = []string{".jpg", ".jpeg", ".png", ".gif", ".webp", ".bmp", ".svg"}
+var imageFileExtensions = []string{".jpg", ".jpeg", ".png", ".gif", ".webp", ".bmp", ".svg", ".tiff", ".tif", ".ico"}
 
 // If file is an image, create link for hotlinking
 func addHotlink(file *models.File) {
@@ -272,7 +511,7 @@ func RequiresClientDecryption(file models.File) bool {
 	if !file.Encryption.IsEncrypted {
 		return false
 	}
-	return file.AwsBucket != ""
+	return !file.IsLocalStorage() || file.Encryption.IsEndToEndEncrypted
 }
 
 // ServeFile subtracts a download allowance and serves the file to the browser
@@ -282,8 +521,7 @@ func ServeFile(file models.File, w http.ResponseWriter, r *http.Request, forceDo
 	database.SaveMetaData(file)
 	logging.AddDownload(&file, r)
 
-	// If file is stored on AWS
-	if file.AwsBucket != "" {
+	if !file.IsLocalStorage() {
 		// We are not setting a download complete status as there is no reliable way to
 		// confirm that the file has been completely downloaded. It expires automatically after 24 hours.
 		downloadstatus.SetDownload(file)
@@ -292,13 +530,20 @@ func ServeFile(file models.File, w http.ResponseWriter, r *http.Request, forceDo
 		return
 	}
 	fileData, size := getFileHandler(file, configuration.Get().DataDir)
+	if file.Encryption.IsEncrypted && !RequiresClientDecryption(file) {
+		if !encryption.IsCorrectKey(file.Encryption, fileData) {
+			w.Write([]byte("Internal error - Error decrypting file, source data might be damaged or an incorrect key has been used"))
+			return
+		}
+	}
 	statusId := downloadstatus.SetDownload(file)
 	writeDownloadHeaders(file, w, forceDownload)
 	if file.Encryption.IsEncrypted && !RequiresClientDecryption(file) {
 		err := encryption.DecryptReader(file.Encryption, fileData, w)
 		if err != nil {
 			w.Write([]byte("Error decrypting file"))
-			panic(err)
+			fmt.Println(err)
+			return
 		}
 	} else {
 		w.Header().Set("Content-Length", strconv.FormatInt(size, 10))
@@ -322,7 +567,7 @@ func writeDownloadHeaders(file models.File, w http.ResponseWriter, forceDownload
 }
 
 func getFileHandler(file models.File, dataDir string) (*os.File, int64) {
-	storageData, err := os.OpenFile(dataDir+"/"+file.SHA256, os.O_RDONLY, 0644)
+	storageData, err := os.OpenFile(dataDir+"/"+file.SHA1, os.O_RDONLY, 0644)
 	helper.Check(err)
 	size, err := helper.GetFileSize(storageData)
 	helper.Check(err)
@@ -331,15 +576,21 @@ func getFileHandler(file models.File, dataDir string) (*os.File, int64) {
 
 // FileExists checks if the file exists locally or in S3
 func FileExists(file models.File, dataDir string) bool {
-	if file.AwsBucket != "" {
-		result, _, err := aws.FileExists(file)
+	if !file.IsLocalStorage() {
+		exists, size, err := aws.FileExists(file)
 		if err != nil {
 			fmt.Println("Warning, cannot check file " + file.Id + ": " + err.Error())
 			return true
 		}
-		return result
+		if !exists {
+			return false
+		}
+		if size == 0 && file.Size != "0 B" {
+			return false
+		}
+		return true
 	}
-	return helper.FileExists(dataDir + "/" + file.SHA256)
+	return helper.FileExists(dataDir + "/" + file.SHA1)
 }
 
 // CleanUp removes expired files from the config and from the filesystem if they are not referenced by other files anymore
@@ -354,7 +605,7 @@ func CleanUp(periodic bool) {
 		if !fileExists || isExpiredFileWithoutDownload(element, timeNow) {
 			deleteFile := true
 			for _, secondLoopElement := range database.GetAllMetadata() {
-				if (element.Id != secondLoopElement.Id) && (element.SHA256 == secondLoopElement.SHA256) {
+				if (element.Id != secondLoopElement.Id) && (element.SHA1 == secondLoopElement.SHA1) {
 					deleteFile = false
 				}
 			}
@@ -371,6 +622,8 @@ func CleanUp(periodic bool) {
 	if wasItemDeleted {
 		CleanUp(false)
 	}
+	cleanOldTempFiles()
+
 	if periodic {
 		go func() {
 			select {
@@ -380,6 +633,38 @@ func CleanUp(periodic bool) {
 		}()
 	}
 	database.RunGarbageCollection()
+}
+
+func cleanOldTempFiles() {
+	tmpfiles, err := os.ReadDir(configuration.Get().DataDir)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	for _, file := range tmpfiles {
+		if isOldTempFile(file) {
+			err = os.Remove(configuration.Get().DataDir + "/" + file.Name())
+			if err != nil {
+				fmt.Println(err)
+			}
+		}
+	}
+}
+
+// Returns true if a file is older than 24 hours and starts with the name upload or chunk
+func isOldTempFile(file os.DirEntry) bool {
+	if file.IsDir() {
+		return false
+	}
+	if !strings.HasPrefix(file.Name(), "upload") && !strings.HasPrefix(file.Name(), "chunk-") {
+		return false
+	}
+	info, err := file.Info()
+	if err != nil {
+		return false
+	}
+	return time.Now().Sub(info.ModTime()) > 24*time.Hour
+
 }
 
 // IsExpiredFile returns true if the file is expired, either due to download count
@@ -398,10 +683,10 @@ func isExpiredFileWithoutDownload(file models.File, timeNow int64) bool {
 
 func deleteSource(file models.File, dataDir string) {
 	var err error
-	if file.AwsBucket != "" {
+	if !file.IsLocalStorage() {
 		_, err = aws.DeleteObject(file)
 	} else {
-		err = os.Remove(dataDir + "/" + file.SHA256)
+		err = os.Remove(dataDir + "/" + file.SHA1)
 	}
 	if err != nil {
 		fmt.Println("Warning, cannot delete file " + file.Id + ": " + err.Error())
