@@ -5,7 +5,6 @@ Main routine
 */
 
 import (
-	"flag"
 	"fmt"
 	"github.com/forceu/gokapi/internal/configuration"
 	"github.com/forceu/gokapi/internal/configuration/cloudconfig"
@@ -13,7 +12,7 @@ import (
 	"github.com/forceu/gokapi/internal/configuration/setup"
 	"github.com/forceu/gokapi/internal/encryption"
 	"github.com/forceu/gokapi/internal/environment"
-	"github.com/forceu/gokapi/internal/helper"
+	"github.com/forceu/gokapi/internal/environment/flagparser"
 	"github.com/forceu/gokapi/internal/logging"
 	"github.com/forceu/gokapi/internal/storage"
 	"github.com/forceu/gokapi/internal/storage/cloudstorage/aws"
@@ -30,15 +29,16 @@ import (
 
 // Version is the current version in readable form.
 // The go generate call below needs to be modified as well
-const Version = "1.5.2"
+const Version = "1.6.2"
 
-//go:generate sh "../../build/setVersionTemplate.sh" "1.5.2"
+//go:generate sh "../../build/setVersionTemplate.sh" "1.6.1"
 //go:generate sh -c "cp \"$(go env GOROOT)/misc/wasm/wasm_exec.js\" ../../internal/webserver/web/static/js/ && echo Copied wasm_exec.js"
-//go:generate sh -c "GOOS=js GOARCH=wasm go build -o ../../internal/webserver/web/main.wasm github.com/forceu/gokapi/cmd/wasmdownloader && echo Compiled WASM module"
+//go:generate sh -c "GOOS=js GOARCH=wasm go build -o ../../internal/webserver/web/main.wasm github.com/forceu/gokapi/cmd/wasmdownloader && echo Compiled Downloader WASM module"
+//go:generate sh -c "GOOS=js GOARCH=wasm go build -o ../../internal/webserver/web/e2e.wasm github.com/forceu/gokapi/cmd/wasme2e && echo Compiled E2E WASM module"
 
 // Main routine that is called on startup
 func main() {
-	passedFlags := parseFlags()
+	passedFlags := flagparser.ParseFlags()
 	showVersion(passedFlags)
 	rand.Seed(time.Now().UnixNano())
 	fmt.Println(logo)
@@ -49,8 +49,7 @@ func main() {
 	encryption.Init(*configuration.Get())
 	authentication.Init(configuration.Get().Authentication)
 	createSsl(passedFlags)
-	initCloudConfig()
-
+	initCloudConfig(passedFlags)
 	go storage.CleanUp(true)
 	logging.AddString("Gokapi started")
 	go webserver.Start()
@@ -69,8 +68,8 @@ func shutdown() {
 }
 
 // Checks for command line arguments that have to be parsed before loading the configuration
-func showVersion(passedFlags flags) {
-	if passedFlags.showVersion {
+func showVersion(passedFlags flagparser.MainFlags) {
+	if passedFlags.ShowVersion {
 		fmt.Println("Gokapi v" + Version)
 		fmt.Println()
 		fmt.Println("Builder: " + environment.Builder)
@@ -116,15 +115,16 @@ func parseBuildSettings(infos []debug.BuildSetting) {
 	}
 }
 
-func initCloudConfig() {
+func initCloudConfig(passedFlags flagparser.MainFlags) {
 	cConfig, ok := cloudconfig.Load()
 	if ok && aws.Init(cConfig.Aws) {
 		fmt.Println("Saving new files to cloud storage")
 		encLevel := configuration.Get().Encryption.Level
-		if encLevel == encryption.FullEncryptionStored || encLevel == encryption.FullEncryptionInput {
+		if (encLevel == encryption.FullEncryptionStored || encLevel == encryption.FullEncryptionInput) && !passedFlags.DisableCorsCheck {
 			ok, err := aws.IsCorsCorrectlySet(cConfig.Aws.Bucket, configuration.Get().ServerUrl)
 			if err != nil {
 				fmt.Println("Warning: Cannot check CORS settings. " + err.Error())
+				fmt.Println("If your provider does not implement the CORS API call and you are certain that it is set correctly, you can disable this check with --disable-cors-check")
 			} else {
 				if !ok {
 					fmt.Println("Warning: CORS settings for bucket " + cConfig.Aws.Bucket + " might not be set correctly. Download might not be possible with encryption.")
@@ -136,38 +136,17 @@ func initCloudConfig() {
 	}
 }
 
-func parseFlags() flags {
-	passedFlags := flag.FlagSet{}
-	versionShortFlag := passedFlags.Bool("v", false, "Show version info")
-	versionLongFlag := passedFlags.Bool("version", false, "Show version info")
-	reconfigureFlag := passedFlags.Bool("reconfigure", false, "Runs setup again to change Gokapi configuration / passwords")
-	createSslFlag := passedFlags.Bool("create-ssl", false, "Creates a new SSL certificate valid for 365 days")
-	err := passedFlags.Parse(os.Args[1:])
-	helper.Check(err)
-	return flags{
-		showVersion: *versionShortFlag || *versionLongFlag,
-		reconfigure: *reconfigureFlag,
-		createSsl:   *createSslFlag,
-	}
-}
-
 // Checks for command line arguments that have to be parsed after loading the configuration
-func reconfigureServer(passedFlags flags) {
-	if passedFlags.reconfigure {
+func reconfigureServer(passedFlags flagparser.MainFlags) {
+	if passedFlags.Reconfigure {
 		setup.RunConfigModification()
 	}
 }
 
-func createSsl(passedFlags flags) {
-	if passedFlags.createSsl {
+func createSsl(passedFlags flagparser.MainFlags) {
+	if passedFlags.CreateSsl {
 		ssl.GenerateIfInvalidCert(configuration.Get().ServerUrl, true)
 	}
-}
-
-type flags struct {
-	showVersion bool
-	reconfigure bool
-	createSsl   bool
 }
 
 var osExit = os.Exit
