@@ -3,6 +3,7 @@
 package main
 
 import (
+	"github.com/forceu/gokapi/internal/environment/flagparser"
 	"github.com/forceu/gokapi/internal/test"
 	"github.com/forceu/gokapi/internal/test/testconfiguration"
 	"os"
@@ -16,33 +17,20 @@ func TestMain(m *testing.M) {
 	os.Exit(exitVal)
 }
 
-func TestParseFlags(t *testing.T) {
-	os.Args = []string{"gokapi", "--version", "--reconfigure", "-create-ssl"}
-	flags := parseFlags()
-	test.IsEqualBool(t, flags.showVersion, true)
-	test.IsEqualBool(t, flags.reconfigure, true)
-	test.IsEqualBool(t, flags.createSsl, true)
-	os.Args = []string{"gokapi", "--reconfigure", "-create-ssl"}
-	flags = parseFlags()
-	test.IsEqualBool(t, flags.showVersion, false)
-	test.IsEqualBool(t, flags.reconfigure, true)
-	test.IsEqualBool(t, flags.createSsl, true)
-}
-
 func TestShowVersion(t *testing.T) {
-	showVersion(flags{})
+	showVersion(flagparser.MainFlags{})
 	osExit = test.ExitCode(t, 0)
-	showVersion(flags{showVersion: true})
+	showVersion(flagparser.MainFlags{ShowVersion: true})
 }
 
 func TestNoResetPw(t *testing.T) {
-	reconfigureServer(flags{})
+	reconfigureServer(flagparser.MainFlags{})
 }
 
 func TestCreateSsl(t *testing.T) {
 	test.FileDoesNotExist(t, "test/ssl.key")
-	createSsl(flags{})
+	createSsl(flagparser.MainFlags{})
 	test.FileDoesNotExist(t, "test/ssl.key")
-	createSsl(flags{createSsl: true})
+	createSsl(flagparser.MainFlags{CreateSsl: true})
 	test.FileExists(t, "test/ssl.key")
 }
