@@ -74,7 +74,7 @@ func TestEncryptionSetup(t *testing.T) {
 	test.IsEqualString(t, config.Encryption.Checksum, "")
 
 	input.EncryptionLevel.Value = "2"
-	input.EncryptionPassword.Value = "testpw"
+	input.EncryptionPassword.Value = "testpw12"
 	formObjects, err = input.toFormObject()
 	test.IsNil(t, err)
 	config, _, err = toConfiguration(&formObjects)
@@ -115,7 +115,7 @@ func TestEncryptionSetup(t *testing.T) {
 	_, ok = database.GetMetaDataById(id)
 	test.IsEqualBool(t, ok, true)
 	configuration.Get().Encryption.Level = 2
-	input.EncryptionPassword.Value = "otherpw"
+	input.EncryptionPassword.Value = "otherpw12"
 	id = testconfiguration.WriteEncryptedFile()
 	_, ok = database.GetMetaDataById(id)
 	test.IsEqualBool(t, ok, true)
@@ -227,14 +227,13 @@ func TestInitialSetup(t *testing.T) {
 }
 
 func TestRunConfigModification(t *testing.T) {
+	statusChannel = make(chan bool, 1)
 	testconfiguration.Create(false)
 	username = ""
 	password = ""
 	finish := make(chan bool)
 	go func() {
-		for !serverStarted {
-			time.Sleep(100 * time.Millisecond)
-		}
+		checkServerStatus(t, true)
 		time.Sleep(500 * time.Millisecond)
 		test.HttpPageResult(t, test.HttpTestConfig{
 			Url:             "http://localhost:53842/setup/start",
@@ -252,15 +251,36 @@ func TestRunConfigModification(t *testing.T) {
 	test.IsEqualInt(t, len(password), 10)
 	isInitialSetup = true
 	<-finish
+	statusChannel = nil
+}
+
+func checkServerStatus(t *testing.T, expected bool) {
+	t.Helper()
+	if statusChannel == nil {
+		t.Fatal("statusChannel is nil")
+	}
+	var result bool
+	select {
+	case result = <-statusChannel:
+
+	case <-time.After(20 * time.Second):
+		t.Fatal("statusChannel timeout after 20 seconds")
+	}
+	if result != expected {
+		if expected == true {
+			t.Fatal("Server was started when it was supposed to be shut down")
+		} else {
+			t.Fatal("Server was shut down when it was supposed to be started")
+		}
+	}
 }
 
 func TestIntegration(t *testing.T) {
+	statusChannel = make(chan bool, 1)
 	testconfiguration.Delete()
 	test.FileDoesNotExist(t, "test/config.json")
 	go RunIfFirstStart()
-	for !serverStarted {
-		time.Sleep(100 * time.Millisecond)
-	}
+	checkServerStatus(t, true)
 
 	test.HttpPageResult(t, test.HttpTestConfig{
 		Url:             "http://localhost:53842/admin",
@@ -298,14 +318,7 @@ func TestIntegration(t *testing.T) {
 		Body:            strings.NewReader(setupValues.toJson()),
 	})
 
-	counter := 0
-	for serverStarted {
-		time.Sleep(100 * time.Millisecond)
-		counter++
-		if counter > 100 {
-			t.Fatal("Unbroken loop")
-		}
-	}
+	checkServerStatus(t, false)
 	test.FileExists(t, "test/config.json")
 	settings := configuration.Get()
 	test.IsEqualInt(t, settings.Authentication.Method, 0)
@@ -333,9 +346,7 @@ func TestIntegration(t *testing.T) {
 	test.FileExists(t, "test/cloudconfig.yml")
 
 	go RunConfigModification()
-	for !serverStarted {
-		time.Sleep(100 * time.Millisecond)
-	}
+	checkServerStatus(t, true)
 
 	username = "test"
 	password = "testpw"
@@ -380,14 +391,8 @@ func TestIntegration(t *testing.T) {
 		Body:            strings.NewReader(setupInput.toJson()),
 	})
 
-	counter = 0
-	for serverStarted {
-		time.Sleep(100 * time.Millisecond)
-		counter++
-		if counter > 100 {
-			t.Fatal("Unbroken loop")
-		}
-	}
+	checkServerStatus(t, false)
+
 	test.FileExists(t, "test/config.json")
 	settings = configuration.Get()
 	test.IsEqualInt(t, settings.Authentication.Method, 2)
@@ -416,9 +421,8 @@ func TestIntegration(t *testing.T) {
 	test.FileDoesNotExist(t, "test/cloudconfig.yml")
 
 	go RunConfigModification()
-	for !serverStarted {
-		time.Sleep(100 * time.Millisecond)
-	}
+	checkServerStatus(t, true)
+
 	username = "test"
 	password = "testpw"
 
@@ -434,14 +438,7 @@ func TestIntegration(t *testing.T) {
 		Body:            strings.NewReader(setupInput.toJson()),
 	})
 
-	counter = 0
-	for serverStarted {
-		time.Sleep(100 * time.Millisecond)
-		counter++
-		if counter > 100 {
-			t.Fatal("Unbroken loop")
-		}
-	}
+	checkServerStatus(t, false)
 
 	test.IsEqualBool(t, settings.PicturesAlwaysLocal, true)
 	test.IsEqualString(t, settings.Authentication.OauthProvider, "provider")
@@ -453,6 +450,7 @@ func TestIntegration(t *testing.T) {
 		test.IsEqualString(t, settings.Authentication.OauthUsers[0], "oatest1")
 		test.IsEqualString(t, settings.Authentication.OauthUsers[1], "oatest2")
 	}
+	statusChannel = nil
 }
 
 type setupValues struct {
@@ -519,19 +517,6 @@ func (s *setupValues) toFormObject() ([]jsonFormObject, error) {
 	return setupResult, err
 }
 
-func TestIsPwLongEnough(t *testing.T) {
-	isInitialSetup = true
-	test.IsEqualBool(t, isPwLongEnough("unc"), false)
-	test.IsEqualBool(t, isPwLongEnough("12345"), false)
-	test.IsEqualBool(t, isPwLongEnough("123456"), true)
-	test.IsEqualBool(t, isPwLongEnough("1234567"), true)
-	isInitialSetup = false
-	test.IsEqualBool(t, isPwLongEnough("unc"), true)
-	test.IsEqualBool(t, isPwLongEnough("12345"), false)
-	test.IsEqualBool(t, isPwLongEnough("123456"), true)
-	test.IsEqualBool(t, isPwLongEnough("1234567"), true)
-}
-
 func createInvalidSetupValues() []setupValues {
 	var result []setupValues
 	input := createInputInternalAuth()
@@ -565,10 +550,6 @@ func createInvalidSetupValues() []setupValues {
 
 	invalidSetup = input
 	invalidSetup.EncryptionLevel.Value = "9"
-	result = append(result, invalidSetup)
-
-	invalidSetup = input
-	invalidSetup.EncryptionLevel.Value = "5" // e2e not implemented yet
 	result = append(result, invalidSetup)
 
 	invalidSetup = input
